Move table protocol to new module
